fix: adjust code up colors by current theme
