+++ src/components/editor/editor.tsx
@@ -102,7 +102,7 @@ const NoteEditor = ({
             Placeholder.configure({
               placeholder:
                 variant !== 'postInput'
-                  ? "Press '/' to commands"
+                  ? "Write or press '/' for commands..."
                   : 'Start writing...'
             })
           ]}
