+++ src/components/editor/selectors/color-selector.tsx
@@ -7,89 +7,12 @@ import {
   PopoverContent,
   PopoverTrigger
 } from '@/components/ui/popover'
-
-export interface BubbleColorMenuItem {
-  name: string
-  color: string
-}
-
-const TEXT_COLORS: BubbleColorMenuItem[] = [
-  {
-    name: 'Default',
-    color: 'var(--novel-black)'
-  },
-  {
-    name: 'Purple',
-    color: '#9333EA'
-  },
-  {
-    name: 'Red',
-    color: '#E00000'
-  },
-  {
-    name: 'Yellow',
-    color: '#EAB308'
-  },
-  {
-    name: 'Blue',
-    color: '#2563EB'
-  },
-  {
-    name: 'Green',
-    color: '#008A00'
-  },
-  {
-    name: 'Orange',
-    color: '#FFA500'
-  },
-  {
-    name: 'Pink',
-    color: '#BA4081'
-  },
-  {
-    name: 'Gray',
-    color: '#A8A29E'
-  }
-]
-
-const HIGHLIGHT_COLORS: BubbleColorMenuItem[] = [
-  {
-    name: 'Default',
-    color: 'var(--novel-highlight-default)'
-  },
-  {
-    name: 'Purple',
-    color: 'var(--novel-highlight-purple)'
-  },
-  {
-    name: 'Red',
-    color: 'var(--novel-highlight-red)'
-  },
-  {
-    name: 'Yellow',
-    color: 'var(--novel-highlight-yellow)'
-  },
-  {
-    name: 'Blue',
-    color: 'var(--novel-highlight-blue)'
-  },
-  {
-    name: 'Green',
-    color: 'var(--novel-highlight-green)'
-  },
-  {
-    name: 'Orange',
-    color: 'var(--novel-highlight-orange)'
-  },
-  {
-    name: 'Pink',
-    color: 'var(--novel-highlight-pink)'
-  },
-  {
-    name: 'Gray',
-    color: 'var(--novel-highlight-gray)'
-  }
-]
+import { useTheme } from 'next-themes'
+import {
+  DARK_TEXT_COLORS,
+  HIGHLIGHT_COLORS,
+  LIGHT_TEXT_COLORS
+} from '../colors'
 
 interface ColorSelectorProps {
   open: boolean
@@ -97,7 +20,10 @@ interface ColorSelectorProps {
 }
 
 export const ColorSelector = ({ open, onOpenChange }: ColorSelectorProps) => {
+  const { theme } = useTheme()
   const { editor } = useEditor()
+
+  const TEXT_COLORS = theme === 'light' ? LIGHT_TEXT_COLORS : DARK_TEXT_COLORS
 
   if (!editor) return null
   const activeColorItem = TEXT_COLORS.find(({ color }) =>
