+++ src/components/rich-text-input/toolbar.tsx
@@ -1,6 +1,5 @@
 'use client'
 import { Toggle } from '@/components/ui/toggle'
-import { type Editor } from '@tiptap/core'
 import { useMemo, useState } from 'react'
 import { LinkSelector } from '../editor/selectors/link-selector'
 import { nodeItems } from '../editor/selectors/node-selector'
@@ -8,11 +7,7 @@ import { textItems } from '../editor/selectors/text-buttons'
 import { useEditorUploadFile } from '../editor/use-image-upload'
 import { Icons } from '../icons'
 import { ToolbarColorSelector } from './toolbar-color-selector'
-
-interface ToolbarProps {
-  editor: Editor | null
-  hasTitleEnabled?: boolean
-}
+import { type ToolbarProps } from '../editor/types'
 
 export const Toolbar = ({ editor, hasTitleEnabled = true }: ToolbarProps) => {
   const [openColor, setOpenColor] = useState(false)
@@ -30,7 +25,7 @@ export const Toolbar = ({ editor, hasTitleEnabled = true }: ToolbarProps) => {
 
   return (
     <div className="flex flex-wrap w-fit max-w-[90vw] rounded-md border border-muted px-2 py-1 bg-background space-x-1">
-      {filteredNodeItems.map((item, index) => (
+      {filteredNodeItems.map(item => (
         <Toggle
           key={item.name}
           className='data-[state="on"]:bg-slate-900'
@@ -84,7 +79,7 @@ export const Toolbar = ({ editor, hasTitleEnabled = true }: ToolbarProps) => {
         <Icons.image size={18} />
       </Toggle>
       <span className="border border-muted" />
-      {textItems.map((item, index) => (
+      {textItems.map(item => (
         <Toggle
           key={item.name}
           className='data-[state="on"]:bg-slate-900'
